- refactor checkpoint
