+++ ConseilJS-softsigner/src/SoftSigner.ts
@@ -1,3 +1,4 @@
+import { Signer } from 'conseiljs';
 
 export class SoftSigner implements Signer {
     //encryptMessage;
