+++ ConseilJS-softsigner/src/utils/CryptoUtils.ts
