+++ ConseilJS-softsigner/src/utils/WrapperWrapper.js
